Clean up and add more comments
