+++ src/mcpdotnet/Protocol/Types/IResourceContents.cs
@@ -1,0 +1,10 @@
+﻿namespace McpDotNet.Protocol.Types;
+
+/// <summary>
+/// Interface for resource contents.
+/// </summary>
+public interface IResourceContents
+{
+    string Uri { get; }
+    string? MimeType { get; }
+}
